♻️ refactor(api/app): update the app to use the new API
